# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Golang REST API wrapper for the MNEE Golang SDK.",
        "title": "MNEE SDK API Wrapper (Go)",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/balance": {
            "get": {
                "description": "Retrieves balances for a comma-separated list of addresses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Get balances for multiple addresses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of addresses",
                        "name": "addresses",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetBalancesSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/balance/{address}": {
            "get": {
                "description": "Retrieves the MNEE balance for a specific wallet address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Balance"
                ],
                "summary": "Get balance for a single address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetBalanceSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
...
        "/transaction/partial-sign": {
            "post": {
                "description": "Builds and signs a transaction *only* with the provided WIFs. Returns hex.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Partial Sign Transaction",
                "parameters": [
                    {
                        "description": "Transfer Parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PartialSignSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
        "/transaction/status/{ticketId}": {
            "get": {
                "description": "Polls the status of a transaction ticket until it is processed or times out.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Poll Ticket Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "ticketId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetTicketSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.GenericFailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.TransferRequest": {
            "type": "object",
            "required": [
                "request",
                "wifs"
            ],
            "properties": {
                "request": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "address",
                            "amount"
                        ],
                        "properties": {
                            "address": {
                                "type": "string",
                                "example": "1G6CB3Ch4zFkPmuhZzEyChQmrQPfi86qk3"
                            },
                            "amount": {
                                "type": "number",
                                "example": 0.1
                            }
                        }
                    }
                },
                "wifs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "L1dRKo...",
                        "K2..."
                    ]
                }
            }
        },
        "models.GenericFailureResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error description"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.GetBalanceSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.BalanceDataDTO"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.GetBalancesSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.BalanceDataDTO"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.GetTicketSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Ticket"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.PartialSignSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.RawTxWrapper"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.RawTxWrapper": {
            "type": "object",
            "properties": {
                "rawTxHex": {
                    "type": "string",
                    "example": "02000000..."
                }
            }
        },
...
        "types.BalanceDataDTO": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amt": {
                    "type": "number"
                },
                "precised": {
                    "type": "number"
                }
            }
        },
...
        "types.Ticket": {
            "type": "object",
            "properties": {
                "action_requested": {
                    "type": "string"
                },
                "callback_secret": {
                    "type": "string"
                },
                "callback_url": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.TicketStatus"
                },
                "tx_hex": {
                    "type": "string"
                },
                "tx_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "types.TicketStatus": {
            "type": "string",
            "enum": [
                "BROADCASTING",
                "SUCCESS"
            ],
            "x-enum-varnames": [
                "BROADCASTING",
                "SUCCESS"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}